Refactored array to include Array class inside array/array.py
